25053: Renamed addresses module to address_handler.
