+++ backend/tests/test_address_handlers.py
@@ -7,7 +7,7 @@
 #
 from . import util
 
-from mora.service.addresses import dar, ean, email, phone, pnumber, text, www
+from mora.service.address_handler import dar, ean, email, phone, pnumber, text, www
 
 
 @util.mock('dawa-addresses.json')
